Imported Upstream version 1.1~rc1

+++ cr-service.c
@@ -19,15 +19,17 @@
 #include "log.h"
 #include "pstree.h"
 #include "cr-service.h"
+#include "cr-service-const.h"
+#include "sd-daemon.h"
 
 unsigned int service_sk_ino = -1;
 
 static int recv_criu_msg(int socket_fd, CriuReq **msg)
 {
-	unsigned char buf[MAX_MSG_SIZE];
+	unsigned char buf[CR_MAX_MSG_SIZE];
 	int len;
 
-	len = read(socket_fd, buf, MAX_MSG_SIZE);
+	len = read(socket_fd, buf, CR_MAX_MSG_SIZE);
 	if (len == -1) {
 		pr_perror("Can't read request");
 		return -1;
@@ -44,7 +46,7 @@ static int recv_criu_msg(int socket_fd, CriuReq **msg)
 
 static int send_criu_msg(int socket_fd, CriuResp *msg)
 {
-	unsigned char buf[MAX_MSG_SIZE];
+	unsigned char buf[CR_MAX_MSG_SIZE];
 	int len;
 
 	len = criu_resp__get_packed_size(msg);
@@ -99,7 +101,7 @@ static int setup_opts_from_req(int sk, CriuOpts *req)
 	char images_dir_path[PATH_MAX];
 
 	if (getsockopt(sk, SOL_SOCKET, SO_PEERCRED, &ids, &ids_len)) {
-		pr_perror("Can't get socket options.");
+		pr_perror("Can't get socket options");
 		return -1;
 	}
 
@@ -132,7 +134,7 @@ static int setup_opts_from_req(int sk, CriuOpts *req)
 
 	log_set_loglevel(req->log_level);
 	if (log_init(opts.output) == -1) {
-		pr_perror("Can't initiate log.");
+		pr_perror("Can't initiate log");
 		return -1;
 	}
 
@@ -166,18 +168,25 @@ static int setup_opts_from_req(int sk, CriuOpts *req)
 static int dump_using_req(int sk, CriuOpts *req)
 {
 	bool success = false;
+	bool self_dump = !req->pid;
 
 	if (setup_opts_from_req(sk, req) == -1) {
 		pr_perror("Arguments treating fail");
 		goto exit;
 	}
 
-	if (cr_dump_tasks(req->pid) == -1)
+	/*
+	 * FIXME -- cr_dump_tasks() may return code from custom
+	 * scripts, that can be positive. However, right now we
+	 * don't have ability to push scripts via RPC, so psitive
+	 * ret values are impossible here.
+	 */
+	if (cr_dump_tasks(req->pid))
 		goto exit;
 
-	if (req->has_leave_running && req->leave_running) {
-		success = true;
+	success = true;
 exit:
+	if (req->leave_running  || !self_dump) {
 		if (send_criu_dump_resp(sk, success, false) == -1) {
 			pr_perror("Can't send response");
 			success = false;
@@ -209,7 +218,8 @@ static int restore_using_req(int sk, CriuOpts *req)
 
 	success = true;
 exit:
-	if (send_criu_restore_resp(sk, success, root_item->pid.real) == -1) {
+	if (send_criu_restore_resp(sk, success,
+				   root_item ? root_item->pid.real : -1) == -1) {
 		pr_perror("Can't send response");
 		success = false;
 	}
@@ -331,53 +341,61 @@ static int restore_sigchld_handler()
 
 int cr_service(bool daemon_mode)
 {
-	int server_fd = -1;
+	int server_fd = -1, n;
 	int child_pid;
 
-	struct sockaddr_un server_addr;
 	struct sockaddr_un client_addr;
-
-	socklen_t server_addr_len;
 	socklen_t client_addr_len;
 
-	server_fd = socket(AF_LOCAL, SOCK_SEQPACKET, 0);
-	if (server_fd == -1) {
-		pr_perror("Can't initialize service socket.");
+	n = sd_listen_fds(0);
+	if (n > 1) {
+		pr_err("Too many file descriptors (%d) recieved", n);
 		goto err;
-	}
+	} else if (n == 1)
+		server_fd = SD_LISTEN_FDS_START + 0;
+	else {
+		struct sockaddr_un server_addr;
+		socklen_t server_addr_len;
 
-	memset(&server_addr, 0, sizeof(server_addr));
-	memset(&client_addr, 0, sizeof(client_addr));
-	server_addr.sun_family = AF_LOCAL;
+		server_fd = socket(AF_LOCAL, SOCK_SEQPACKET, 0);
+		if (server_fd == -1) {
+			pr_perror("Can't initialize service socket");
+			goto err;
+		}
 
-	if (opts.addr == NULL)
-		opts.addr = CR_DEFAULT_SERVICE_ADDRESS;
+		memset(&server_addr, 0, sizeof(server_addr));
+		memset(&client_addr, 0, sizeof(client_addr));
+		server_addr.sun_family = AF_LOCAL;
 
-	strcpy(server_addr.sun_path, opts.addr);
+		if (opts.addr == NULL)
+			opts.addr = CR_DEFAULT_SERVICE_ADDRESS;
 
-	server_addr_len = strlen(server_addr.sun_path)
-			+ sizeof(server_addr.sun_family);
-	client_addr_len = sizeof(client_addr);
+		strcpy(server_addr.sun_path, opts.addr);
 
-	unlink(server_addr.sun_path);
+		server_addr_len = strlen(server_addr.sun_path)
+				+ sizeof(server_addr.sun_family);
+		client_addr_len = sizeof(client_addr);
 
-	if (bind(server_fd, (struct sockaddr *) &server_addr,
-					server_addr_len) == -1) {
-		pr_perror("Can't bind.");
-		goto err;
-	}
+		unlink(server_addr.sun_path);
 
-	pr_info("The service socket is bound to %s\n", server_addr.sun_path);
+		if (bind(server_fd, (struct sockaddr *) &server_addr,
+						server_addr_len) == -1) {
+			pr_perror("Can't bind");
+			goto err;
+		}
 
-	/* change service socket permissions, so anyone can connect to it */
-	if (chmod(server_addr.sun_path, 0666)) {
-		pr_perror("Can't change permissions of the service socket.");
-		goto err;
-	}
+		pr_info("The service socket is bound to %s\n", server_addr.sun_path);
 
-	if (listen(server_fd, 16) == -1) {
-		pr_perror("Can't listen for socket connections.");
-		goto err;
+		/* change service socket permissions, so anyone can connect to it */
+		if (chmod(server_addr.sun_path, 0666)) {
+			pr_perror("Can't change permissions of the service socket");
+			goto err;
+		}
+
+		if (listen(server_fd, 16) == -1) {
+			pr_perror("Can't listen for socket connections");
+			goto err;
+		}
 	}
 
 	if (daemon_mode) {
@@ -390,7 +408,7 @@ int cr_service(bool daemon_mode)
 	if (opts.pidfile) {
 		if (write_pidfile(getpid()) == -1) {
 			pr_perror("Can't write pidfile");
-			return -1;
+			goto err;
 		}
 	}
 
@@ -404,7 +422,7 @@ int cr_service(bool daemon_mode)
 
 		sk = accept(server_fd, &client_addr, &client_addr_len);
 		if (sk == -1) {
-			pr_perror("Can't accept connection.");
+			pr_perror("Can't accept connection");
 			goto err;
 		}
 
@@ -419,7 +437,7 @@ int cr_service(bool daemon_mode)
 			close(server_fd);
 			ret = cr_service_work(sk);
 			close(sk);
-			exit(ret);
+			exit(ret != 0);
 		}
 
 		if (child_pid < 0)
